pyproject.toml
[project]
name = "python-meteor-test"
version = "0.1.0"
description = ""
authors = [
    {name = "renemazuela",email = "rjmazuela@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.79.0,<2.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "setuptools (>=80.8.0,<81.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "markdown (>=3.8,<4.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
